FF4 update: Ensure the extension is unpacked.
As of FF4 extensions by default are left in their xpi; but the current
code expects to load scripts from the components directory. In the long
run the fix is to move those scripts into chrome/content/ and to load them
via the "chrome://<extension>/content/<file>" url, but until then we force
the unpacking.

See also https://bugzilla.mozilla.org/show_bug.cgi?id=533038 and
https://developer.mozilla.org/en/Firefox_4_for_developers#XPI_unpacking .